analytics/requirements.txt
ibis-framework[duckdb]==9.5.0
matplotlib==3.10.0